reorganize to register all singledispatches at anndata import
