+++ src/anndata/_core/anndata.py
@@ -42,6 +42,7 @@ from .index import _normalize_indices, _subset, get_vector
 from .raw import Raw
 from .sparse_dataset import BaseCompressedSparseDataset, sparse_dataset
 from .storage import coerce_array
+from .xarray import Dataset2D
 from .views import (
     DictView,
     _resolve_idxs,
@@ -2078,6 +2079,12 @@ class AnnData(metaclass=utils.DeprecationMixinMeta):
         getattr(self, a).drop(keys, axis=1, inplace=True)
         return values
 
+@AnnData._remove_unused_categories.register(Dataset2D)
+@staticmethod
+def _remove_unused_categories_xr(
+    df_full: Dataset2D, df_sub: Dataset2D, uns: dict[str, Any]
+):
+    pass  # this is handled automatically by the categorical arrays themselves i.e., they dedup upon access.
 
 def _check_2d_shape(X):
     """\
