+++ src/anndata/_core/file_backing.py
@@ -9,6 +9,7 @@ from typing import TYPE_CHECKING
 import h5py
 
 from ..compat import AwkArray, DaskArray, ZarrArray, ZarrGroup
+from .xarray import Dataset2D
 from .sparse_dataset import BaseCompressedSparseDataset
 
 if TYPE_CHECKING:
@@ -160,6 +161,10 @@ def _(x: AwkArray, *, copy: bool = False):
         return _copy(x)
     else:
         return x
+
+@to_memory.register(Dataset2D)
+def _(x: Dataset2D, copy: bool = False):
+    return x.to_memory(copy=copy)
 
 
 @singledispatch
