+++ src/anndata/_core/views.py
@@ -22,6 +22,7 @@ from ..compat import (
     DaskArray,
     ZappyArray,
 )
+from .xarray import Dataset2D
 from .access import ElementRef
 
 if TYPE_CHECKING:
@@ -360,6 +361,10 @@ def as_view_cupy_csr(mtx, view_args):
 @as_view.register(CupyCSCMatrix)
 def as_view_cupy_csc(mtx, view_args):
     return CupySparseCSCView(mtx, view_args=view_args)
+
+@as_view.register(Dataset2D)
+def _(a: Dataset2D, view_args):
+    return a
 
 
 try:
